feat(svg&nom): Renomage du composant du bouton et mis en composant du svg
